Unify FunctionSource in Map and Reduce
Note that this is not entirely for free.  The function aliases (example: where we alias 'max' to 'amax' for `FunctionSource.np`) now come from one source, which may or may not be desirable.
