+++ docs/source/_static/data_processing_examples/iss_pipeline.py
@@ -8,7 +8,7 @@ import os
 import starfish
 from starfish.image import ApplyTransform, Filter, LearnTransform, Segment
 from starfish.spots import FindSpots, DecodeSpots, AssignTargets
-from starfish.types import Axes
+from starfish.types import Axes, FunctionSource
 
 test = os.getenv("TESTING") is not None
 
@@ -20,7 +20,7 @@ def iss_pipeline(fov, codebook):
     learn_translation = LearnTransform.Translation(reference_stack=fov.get_image('dots'),
                                                    axes=Axes.ROUND, upsampling=100)
     max_projector = Filter.Reduce(
-        (Axes.CH, Axes.ZPLANE), func="max", module=Filter.Reduce.FunctionSource.np)
+        (Axes.CH, Axes.ZPLANE), func="max", module=FunctionSource.np)
     transforms_list = learn_translation.run(max_projector.run(primary_image))
     warp = ApplyTransform.Warp()
     registered = warp.run(primary_image, transforms_list=transforms_list,  in_place=False, verbose=True)
