+++ starfish/core/types/__init__.py
@@ -14,6 +14,7 @@ from ._constants import (
     TransformType
 )
 from ._decoded_spots import DecodedSpots
+from ._functionsource import FunctionSource
 from ._spot_attributes import SpotAttributes
 from ._spot_finding_results import SpotFindingResults
